src/raspi-base/radio-daemon/requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cchardet==2.1.4
certifi==2018.8.24
cffi==1.13.2
chardet==3.0.4
Click==7.0
decorator==4.4.1
future==0.18.2
geocoder==1.38.1
idna==2.6
multidict==4.5.2
pycares==3.0.0
pycparser==2.19
python-daemon==2.2.4
python-dotenv==0.11.0
python-http-client==3.2.4
ratelim==0.1.6
ratelimitingfilter==1.2
requests==2.21.0
rpi-rfm69==0.2.1b0
RPi.GPIO==0.7.0
sendgrid==6.1.1
six==1.12.0
spidev==3.4
ssh-import-id==5.7
urllib3==1.24.1
virtualenv==16.7.3
yarl==1.3.0
